ai-roadmap-builder/backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Validation & config
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# LLM
google-generativeai>=0.3.0

# HTTP client (optional: external APIs)
httpx>=0.26.0
